requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==41.0.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dj-static==0.0.6
Django==4.1.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.14.0
django-csp==3.7
django-csv-utils==1.0.2
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==22.1
django-health-check==3.17.0
django-otp==1.1.4
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
executing==1.2.0
fastjsonschema==2.16.3
fqdn==1.5.1
gevent==23.9.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.12.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
lazy-object-proxy==1.9.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
model-bakery==1.10.1
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.0
PyJWT==2.6.0
pylint==2.16.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Send2Trash==1.8.0
six==1.16.0
smart-getenv==1.1.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.5.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
wrapt==1.15.0
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
